+++ open_crawler/api/factory.py
@@ -1,0 +1,34 @@
+import os
+
+from fastapi import FastAPI
+from fastapi.middleware.cors import CORSMiddleware
+
+from api.crawls_router import crawls_router
+from api.websites_router import websites_router
+
+
+def create_api_app() -> FastAPI:
+    """
+    Create and configure the FastAPI application.
+    """
+    api_app = FastAPI(
+        title="Asynchronous tasks processing with Celery and RabbitMQ",
+        description="Sample FastAPI Application to demonstrate Event "
+        "driven architecture with Celery and RabbitMQ",
+        version="1.0.0",
+    )
+
+    # Configure CORS for non-production modes
+    deployment_mode = os.environ.get("MODE", "production")
+    if deployment_mode != "production":
+        api_app.add_middleware(
+            CORSMiddleware,
+            allow_origins=["*"],
+            allow_credentials=True,
+            allow_methods=["*"],
+            allow_headers=["*"],
+        )
+
+    api_app.include_router(websites_router)
+    api_app.include_router(crawls_router)
+    return api_app
